update inmem backend with tests
